Support for more grip types

+++ Sources/Framework/Startup.ahk
@@ -204,30 +204,32 @@ loadSimulatorConfiguration() {
 		writeMultiMap(kUserHomeDirectory . "Diagnostics\Usage.stat", usage)
 	}
 
-	if !FileExist(kUserHomeDirectory . "Diagnostics\UPLOAD")
-		FileAppend(A_Now, kUserHomeDirectory . "Diagnostics\UPLOAD")
-
-	lastModified := FileGetTime(kUserHomeDirectory . "Diagnostics\UPLOAD", "M")
-
-	lastModified := DateAdd(lastModified, 1, "Days")
-
-	if (lastModified < A_Now) {
-		try {
-			deleteFile(kUserHomeDirectory . "Diagnostics\UPLOAD")
-
+	if (StrSplit(A_ScriptName, ".")[1] = "Simulator Startup") {
+		if !FileExist(kUserHomeDirectory . "Diagnostics\UPLOAD")
 			FileAppend(A_Now, kUserHomeDirectory . "Diagnostics\UPLOAD")
+
+		lastModified := FileGetTime(kUserHomeDirectory . "Diagnostics\UPLOAD", "M")
+
+		lastModified := DateAdd(lastModified, 1, "Days")
+
+		if (lastModified < A_Now) {
+			try {
+				deleteFile(kUserHomeDirectory . "Diagnostics\UPLOAD")
+
+				FileAppend(A_Now, kUserHomeDirectory . "Diagnostics\UPLOAD")
+			}
+
+			Task.startTask(() {
+				ID := StrSplit(FileRead(kUserConfigDirectory . "ID"), "`n", "`r")[1]
+
+				fileName := (ID . "." . A_Now . ".log")
+
+				if ftpUpload("87.177.159.148", "SimulatorController", "Sc-1234567890-Sc", kUserHomeDirectory . "Diagnostics\Critical.log", "Diagnostics-Uploads/" . fileName)
+					deleteFile(kUserHomeDirectory . "Diagnostics\Critical.log")
+
+				ftpUpload("87.177.159.148", "SimulatorController", "Sc-1234567890-Sc", kUserHomeDirectory . "Diagnostics\Usage.stat", "Diagnostics-Uploads/" . ID . ".Usage.stat")
+			}, 10000, kLowPriority)
 		}
-
-		Task.startTask(() {
-			ID := StrSplit(FileRead(kUserConfigDirectory . "ID"), "`n", "`r")[1]
-
-			fileName := (ID . "." . A_Now . ".log")
-
-			if ftpUpload("87.177.159.148", "SimulatorController", "Sc-1234567890-Sc", kUserHomeDirectory . "Diagnostics\Critical.log", "Diagnostics-Uploads/" . fileName)
-				deleteFile(kUserHomeDirectory . "Diagnostics\Critical.log")
-
-			ftpUpload("87.177.159.148", "SimulatorController", "Sc-1234567890-Sc", kUserHomeDirectory . "Diagnostics\Usage.stat", "Diagnostics-Uploads/" . ID . ".Usage.stat")
-		}, 10000, kLowPriority)
 	}
 
 	if kLogStartup
